requirements.txt
requests==2.31.0
weasyprint==61.2
jinja2==3.1.2
python-dotenv==1.0.0
